backend_system/ai_agent_platform/security.py
```python
import hashlib
import os
from cryptography.fernet import Fernet

class Security:

    def __init__(self):
        self.key = os.getenv('SECRET_KEY')
        self.cipher_suite = Fernet(self.key)

    def encrypt_data(self, data):
        """
        Encrypts the data
        """
        encrypted_data = self.cipher_suite.encrypt(data.encode())
        return encrypted_data

    def decrypt_data(self, encrypted_data):
        """
        Decrypts the data
        """
        decrypted_data = self.cipher_suite.decrypt(encrypted_data).decode()
        return decrypted_data

    def hash_password(self, password):
        """
        Hashes the password using SHA256
        """
        hashed_password = hashlib.sha256(password.encode()).hexdigest()
        return hashed_password

    def verify_password(self, hashed_password, user_password):
        """
        Verifies the password
        """
        return hashed_password == hashlib.sha256(user_password.encode()).hexdigest()

    def audit_security(self):
        """
        Conducts a security audit
        """
        # This is a placeholder function. In a real-world application, this would involve
        # checking for vulnerabilities in the system.
        pass

    def incident_response(self):
        """
        Handles a security incident
        """
        # This is a placeholder function. In a real-world application, this would involve
        # steps to respond to a security incident, such as notifying users, resetting passwords,
        # and investigating the cause.
        pass
```